requirements/common.txt
alembic==0.9.9
bleach==2.1.3
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
dominate==2.3.1
Faker==0.8.16
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.1
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-MySQLdb==0.2.0
Flask-PageDown==0.2.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
ForgeryPy==0.1
gunicorn==19.9.0
html5lib==1.0.1
httpie==0.9.9
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
Pygments==2.2.0
python-dateutil==2.7.3
python-editor==1.0.3
redis==2.10.6
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.8
text-unidecode==1.2
urllib3==1.23
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.1
